+++ eventium-core/src/Eventium.hs
@@ -11,7 +11,8 @@ import Eventium.EventPublisher as X
 import Eventium.EventSubscription as X
 import Eventium.ProcessManager as X
 import Eventium.Projection as X
-import Eventium.ProjectionCache.Types as X
+import Eventium.ProjectionCache.Cache as X
+import Eventium.ReadModel as X
 import Eventium.Store.Class as X
 import Eventium.TypeEmbedding as X
 import Eventium.UUID as X